requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
affine=2.4.0=pypi_0
albucore=0.0.23=pypi_0
albumentations=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
aom=3.5.0=h27087fc_0
attrs=25.1.0=pypi_0
blas=1.0=mkl
brotli-python=1.1.0=py39hf88036b_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.12.14=hbcca054_0
cachetools=5.5.0=pypi_0
cairo=1.16.0=h35add3b_1015
certifi=2024.12.14=pyhd8ed1ab_0
cffi=1.17.1=py39h15c3d72_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
click=8.1.8=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
contourpy=1.3.0=pypi_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.6.77=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.6.77=0
cuda-nsight-compute=12.4.1=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.6.77=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.6.80=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.6.80=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.6.77=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-version=12.6=3
cuda-visual-tools=11.6.1=0
cycler=0.12.1=pypi_0
dav1d=1.2.1=hd590300_0
einops=0.8.1=pypi_0
eval-type-backport=0.2.2=pypi_0
ffmpeg=6.0.0=gpl_hdbbbd96_103
filelock=3.18.0=pypi_0
font-ttf-dejavu-sans-mono=2.37=0
font-ttf-inconsolata=2.000=0
font-ttf-source-code-pro=2.030=0
font-ttf-ubuntu=0.83=0
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.56.0=pypi_0
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h36c2ea0_0
fsspec=2025.3.0=pypi_0
gds-tools=1.11.1.6=0
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
gmp=6.3.0=hac33072_2
gnutls=3.7.9=hb077bed_0
graphite2=1.3.13=h59595ed_1003
h2=4.1.0=pyhd8ed1ab_1
harfbuzz=7.3.0=hdb3a94d_0
hpack=4.0.0=pyhd8ed1ab_1
huggingface-hub=0.29.3=pypi_0
hyperframe=6.0.1=pyhd8ed1ab_1
icu=72.1=hcb278e6_0
idna=3.10=pyhd8ed1ab_1
imageio=2.37.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
joblib=1.4.2=pypi_0
jpeg=9e=h0b41bf4_3
kiwisolver=1.4.7=pypi_0
lame=3.100=h166bdaf_1003
lazy-loader=0.4=pypi_0
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.43=h712a8e2_2
lerc=4.0.0=h27087fc_0
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libass=0.17.1=hc9aadba_0
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.11.1.6=0
libcufile-dev=1.11.1.6=0
libcurand=10.3.7.77=0
libcurand-dev=10.3.7.77=0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.17=h0b41bf4_0
libdrm=2.4.124=hb9d3cd8_0
libexpat=2.6.4=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libglib=2.82.2=h2ff4ddf_0
libgomp=14.2.0=h77fa898_1
libiconv=1.17=hd590300_2
libidn2=2.3.7=hd590300_0
liblapack=3.9.0=16_linux64_mkl
liblzma=5.6.3=hb9d3cd8_1
liblzma-devel=5.6.3=hb9d3cd8_1
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libopus=1.3.1=h7f98852_1
libpciaccess=0.18=hd590300_0
libpng=1.6.45=h943b412_0
libsqlite=3.47.2=hee588c1_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtasn1=4.19.0=h166bdaf_0
libtiff=4.5.0=h6adf6a1_2
libunistring=0.9.10=h7f98852_0
libuuid=2.38.1=h0b41bf4_0
libva=2.18.0=h0b41bf4_0
libvpx=1.13.1=h59595ed_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.13=h7f98852_1004
libxcrypt=4.4.36=hd590300_1
libxml2=2.11.5=h0d562d8_0
libzlib=1.3.1=hb9d3cd8_2
matplotlib=3.9.4=pypi_0
mkl=2022.1.0=hc2b9512_224
ncurses=6.5=h2d0b736_2
nettle=3.9.1=h7ab15ed_0
networkx=3.2.1=pypi_0
nsight-compute=2024.1.1.4=0
numpy=1.26.4=pypi_0
nvidia-ml-py=12.560.30=pypi_0
nvitop=1.4.1=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openh264=2.3.1=hcb278e6_2
openjpeg=2.5.0=hfec8fc6_2
openssl=3.4.0=h7b32b05_1
p11-kit=0.24.1=hc5aa10d_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pcre2=10.44=hba22ea6_2
pillow=9.4.0=py39h2320bf1_1
pip=24.3.1=pyh8b19718_2
pixman=0.44.2=h29eaf8c_0
psutil=6.1.1=pypi_0
pthread-stubs=0.4=hb9d3cd8_1002
pycparser=2.22=pyh29332c3_1
pydantic=2.11.1=pypi_0
pydantic-core=2.33.0=pypi_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=pyha55dd90_7
python=3.9.21=h9c0c6dc_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.9=5_cp39
pytorch=1.13.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
qudida=0.0.4=pypi_0
rasterio=1.4.3=pypi_0
readline=8.2=h8228510_1
requests=2.32.3=pyhd8ed1ab_1
safetensors=0.5.3=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
setuptools=75.8.0=pyhff2d567_0
shapely=2.0.7=pypi_0
simsimd=6.2.1=pypi_0
six=1.17.0=pypi_0
stringzilla=3.12.3=pypi_0
svt-av1=1.6.0=h59595ed_0
termcolor=2.5.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.8.30=pypi_0
timm=1.0.15=pypi_0
tk=8.6.13=noxft_h4845f30_101
torchaudio=0.13.1=py39_cu116
torchvision=0.14.1=py39_cu116
tqdm=4.67.1=pypi_0
typing-inspection=0.4.0=pypi_0
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025.1=pypi_0
urllib3=2.3.0=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xorg-fixesproto=5.0=hb9d3cd8_1003
xorg-kbproto=1.0.7=hb9d3cd8_1003
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.5=he73a12e_0
xorg-libx11=1.8.4=h0b41bf4_0
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=hb9d3cd8_1003
xorg-xextproto=7.3.0=hb9d3cd8_1004
xorg-xproto=7.0.31=hb9d3cd8_1008
xz=5.6.3=hbcc6ac9_1
xz-gpl-tools=5.6.3=hbcc6ac9_1
xz-tools=5.6.3=hb9d3cd8_1
zipp=3.21.0=pypi_0
zlib=1.3.1=hb9d3cd8_2
zstandard=0.23.0=py39h08a7858_1
zstd=1.5.6=ha6fb4c9_0
